engines/compliance/requirements.txt
# Compliance Engine Requirements

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# PDF Generation
reportlab>=4.0.0

# Database (Optional)
psycopg2-binary>=2.9.0  # PostgreSQL
boto3>=1.28.0  # DynamoDB
kubernetes>=28.1.0  # K8s Job creation

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0


# ── OpenTelemetry instrumentation ────────────────────────────────────────────
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0
opentelemetry-instrumentation-psycopg2>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0
